feat: create SearchInput component
