+++ src/components/SearchInput/index.ts
@@ -1,0 +1,2 @@
+import SearchInput from "./SearchInput";
+export { SearchInput };
